change default new tab and private new tab structure this allow us to have a better view of the front-end architecture and split interests between files
