+++ components/brave_new_tab_ui/BUILD.gn
@@ -2,26 +2,29 @@ import("//brave/components/common/typescript.gni")
 
 transpile_web_ui("brave_new_tab_ui") {
   inputs = [
-    "brave_new_tab.html",
-    "brave_new_tab.tsx",
-    "storage.ts",
-    "store.ts",
     "actions/new_tab_actions.ts",
-    "components/app.tsx",
-    "components/block.tsx",
-    "components/footerInfo.tsx",
-    "components/siteRemovalNotification.tsx",
-    "components/stats.tsx",
+    "components/newTab/block.tsx",
+    "components/newTab/footerInfo.tsx",
+    "components/newTab/index.tsx",
+    "components/newTab/siteRemovalNotification.tsx",
+    "components/newTab/stats.tsx",
     "components/privateTab/index.tsx",
     "components/privateTab/privateTab.tsx",
     "components/privateTab/qwantTab.tsx",
     "components/privateTab/qwantTorTab.tsx",
+    "components/privateTab/torContent.tsx",
     "components/privateTab/torTab.tsx",
+    "components/app.tsx",
     "constants/backgrounds.ts",
     "constants/new_tab_types.ts",
     "constants/theme.ts",
     "reducers/index.ts",
     "reducers/new_tab_reducer.tsx",
+    "api.ts",
+    "brave_new_tab.html",
+    "brave_new_tab.tsx",
+    "storage.ts",
+    "store.ts"
   ]
 
   entry_points = [
