+++ components/brave_new_tab_ui/brave_new_tab.tsx
@@ -17,6 +17,13 @@ import App from './components/app'
 import store from './store'
 import * as newTabActions from './actions/new_tab_actions'
 
+// Fonts
+import '../fonts/poppins.css'
+import '../fonts/muli.css'
+
+// TODO: remove
+import '../styles/newtab.less'
+
 window.cr.define('brave_new_tab', function () {
   'use strict'
