+++ components/brave_new_tab_ui/components/newTab/block.tsx
@@ -17,7 +17,7 @@ import {
 } from 'react-dnd'
 
 // Utils
-import { cx } from '../../common/classSet'
+import { cx } from '../../../common/classSet'
 
 const Types = {
   BLOCK: 'block'
