+++ components/brave_new_tab_ui/components/newTab/footerInfo.tsx
@@ -3,7 +3,7 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
-import { getLocale } from '../../common/locale'
+import { getLocale } from '../../../common/locale'
 
 interface Props {
   backgroundImage?: NewTab.Image
