+++ components/brave_new_tab_ui/components/newTab/siteRemovalNotification.tsx
