+++ components/brave_new_tab_ui/components/privateTab/qwantTab.tsx
@@ -15,7 +15,7 @@ import {
   TorImage,
   Separator,
   FakeButton
-} from 'brave-ui/features/newTab'
+} from 'brave-ui/features/newTab/private'
 import TorContent from './torContent'
 // Helpers
 import { getLocale } from '../../../common/locale'
