+++ components/brave_new_tab_ui/components/privateTab/qwantTorTab.tsx
@@ -17,7 +17,7 @@ import {
   TorLockImage,
   Separator,
   FakeButton
-} from 'brave-ui/features/newTab'
+} from 'brave-ui/features/newTab/private'
 
 // Helpers
 import { getLocale } from '../../../common/locale'
