+++ components/brave_new_tab_ui/components/privateTab/torContent.tsx
@@ -10,7 +10,7 @@ import {
   SubTitle,
   Text,
   TorLockImage
-} from 'brave-ui/features/newTab'
+} from 'brave-ui/features/newTab/private'
 // Helpers
 import { getLocale } from '../../../common/locale'
